Добавлена задача №3 (исправление ощибок)

+++ src/PrimeNumber/BruteForceJob/Program.cs
@@ -1,13 +1,13 @@
-﻿const long end = 3000000L;
+﻿const int end = 3000000;
 
 Console.WriteLine(312471072265);
 
-bool IsPrime(long n)
+bool IsPrime(int n)
 {
     if (n == 2) return true;
     if (n % 2 == 0) return false;
 
-    long i = 3;
+    int i = 3;
     while (i * i <= n)
     {
         if (n % i == 0) return false;
@@ -18,7 +18,7 @@ bool IsPrime(long n)
 }
 
 long result = 0;
-long i = 2;
+int i = 2;
 
 while (i <= end)
 {
